pyproject.toml
[project]
name = "fusiqc"
version = "0.1.0"
description = "Browser-based QC for fUSI-BIDS datasets."
readme = "README.md"
authors = [
    { name = "Samuel Le Meur-Diebolt", email = "samuel@diebolt.io" }
]
license = "BSD-3-Clause"
requires-python = ">=3.11,<3.14"
dependencies = [
    "confusius==0.0.1a21",
    "pandas>=2.2.3",
    "pybids>=0.18.1",
    "rich>=14.2.0",
]

[project.scripts]
fusiqc = "fusiqc:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.8.0",
    "ty>=0.0.10,<0.1.0",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"
